Fixed loading workflows by WorkflowInstanceFilter with list of states(Now we are restricting to load instances from DB and memory at the same time). Added tests throwing exception if filter states for DB and memory are mixed

+++ projects/copper-jmx-interface/src/main/java/org/copperengine/management/ProcessingEngineMXBean.java
@@ -82,5 +82,5 @@ public interface ProcessingEngineMXBean {
     
     public List<WorkflowInfo> queryWorkflowInstances(WorkflowInstanceFilter filter);
 
-    public int countWorkflowInstances(WorkflowInstanceFilter filter);
+    public long countWorkflowInstances(WorkflowInstanceFilter filter);
 }
